CLN: Group tests for apply/agg/transform with a string argument
